requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_py-xgboost-mutex=2.0=cpu_0
absl-py=0.15.0=pypi_0
aiohttp=3.7.4.post0=pypi_0
altair=4.1.0=pypi_0
argon2-cffi=21.1.0=pypi_0
astor=0.8.1=pypi_0
async-timeout=3.0.1=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
base58=2.1.0=pypi_0
beautifulsoup4=4.10.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blinker=1.4=pypi_0
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.9.30=h06a4308_1
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py39h06a4308_0
cffi=1.15.0=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.7=pypi_0
click=7.1.2=pypi_0
configparser=5.0.2=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.10.0=py39h06a4308_0
datasets=1.14.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.4.1=py39h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
dill=0.3.4=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.3=py39h06a4308_0
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
filelock=3.3.1=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.10.1=pypi_0
gdown=4.2.0=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.24=pypi_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-api-core=2.1.1=pypi_0
google-api-python-client=2.27.0=pypi_0
google-auth=2.3.1=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
googleapis-common-protos=1.53.0=pypi_0
grpcio=1.41.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
httplib2=0.20.1=pypi_0
huggingface-hub=0.0.19=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pypi_0
importlib-metadata=4.8.1=py39h06a4308_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=6.4.1=py39h06a4308_1
ipython=7.27.0=py39hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pypi_0
jedi=0.18.0=py39h06a4308_1
jinja2=3.0.2=pypi_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonschema=4.1.2=pypi_0
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_core=4.8.1=py39h06a4308_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.3.1=py39h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxgboost=1.3.3=h2531618_0
libxml2=2.9.12=h03d6c58_0
lightgbm=3.2.1=py39h295c915_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.3=py39h06a4308_0
matplotlib-base=3.4.3=py39hbbc1b5f_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=pypi_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=5.2.0=pypi_0
multiprocess=0.70.12.2=pypi_0
munkres=1.1.4=py_0
nbclient=0.5.4=pypi_0
nbconvert=6.2.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
nltk=3.6.5=pyhd3eb1b0_0
notebook=6.4.5=pypi_0
numexpr=2.7.3=py39h22e1b3c_1
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
oauth2client=4.1.3=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pypi_0
pandarallel=1.5.4=pypi_0
pandas=1.3.3=py39h8c16a72_0
pandocfilters=1.5.0=pypi_0
parso=0.8.2=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
plotly=5.1.0=pyhd3eb1b0_0
prometheus-client=0.11.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.0=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-xgboost=1.3.3=py39h06a4308_0
pyarrow=5.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pydeck=0.7.1=pypi_0
pydrive=1.3.1=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyrsistent=0.18.0=pypi_0
pysocks=1.7.1=pypi_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.2.1=py39h295c915_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
regex=2021.8.3=py39h7f8727e_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
sacremoses=0.0.46=pypi_0
scikit-learn=0.24.2=py39ha9443f7_0
scipy=1.7.1=py39h292c36d_2
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.96=pypi_0
sentry-sdk=1.4.3=pypi_0
seqeval=1.2.2=pypi_0
setuptools=58.0.4=py39h06a4308_0
shortuuid=1.0.1=pypi_0
simpletransformers=0.62.2=pypi_0
sip=4.19.13=py39h2531618_0
six=1.15.0=py39h06a4308_0
smmap=5.0.0=pypi_0
soupsieve=2.2.1=pypi_0
sqlite=3.36.0=hc218d9a_0
streamlit=1.1.0=pypi_0
subprocess32=3.5.4=pypi_0
tenacity=8.0.1=py39h06a4308_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.11.1=pypi_0
torchaudio=0.10.0=py39_cu113
torchvision=0.11.1=py39_cu113
tornado=6.1=py39h27cfd23_0
tqdm=4.62.2=pyhd3eb1b0_1
traitlets=5.1.0=pyhd3eb1b0_0
transformers=4.11.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021.5=pypi_0
tzlocal=4.0.1=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.7=pypi_0
validators=0.18.2=pypi_0
wandb=0.12.5=pypi_0
watchdog=2.1.6=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=pypi_0
xgboost=1.3.3=py39h06a4308_0
xxhash=2.0.2=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.0=pypi_0
yaspin=2.1.0=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
